Safari crashed and lost all tabs, after unlocking sleeping device https://bugs.webkit.org/show_bug.cgi?id=223832 <rdar://75555287>
Reviewed by David Kilzer.

A speculative fix for this crash. A possible explanation from the crash log suggests that the protectedThis
could be elided because of compiler optimization given it is not used in the block. To prevent such optimization,
protectedThis is therefore used explicitly in the block.

* UIProcess/Cocoa/SOAuthorization/SOAuthorizationSession.mm:
(WebKit::SOAuthorizationSession::dismissViewController):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275197 268f45cc-cd09-0410-ab3c-d52691b4dbfc
Canonical link: https://commits.webkit.org/235894@main