[iOS] Two taps required to view <select> options on Square Checkout https://bugs.webkit.org/show_bug.cgi?id=223933 <rdar://problem/74828029>
Reviewed by Wenson Hsieh.

Source/WebKit:

Some sites, such as Square Checkout, wrap all the <option>s in a
<select> element in a single <optgroup>. With the new context menu
appearance for <select> on iOS, this means that users have to tap twice
to view the options on some sites (once to bring up the menu and once
to expand the grouped menu).

To fix, promote the the contents of the grouped submenu to the root menu
if all the options are contained within a single submenu.

Test: fast/forms/ios/form-control-refresh/select/all-options-in-single-optgroup.html

* UIProcess/ios/forms/WKFormSelectPicker.mm:
(-[WKSelectPicker createMenu]):

LayoutTests:

Added a layout test to exercise the new codepath when a <select>
has a single <optgroup> containing all of the <option>s.

* fast/forms/ios/form-control-refresh/select/all-options-in-single-optgroup-expected.txt: Added.
* fast/forms/ios/form-control-refresh/select/all-options-in-single-optgroup.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275216 268f45cc-cd09-0410-ab3c-d52691b4dbfc
Canonical link: https://commits.webkit.org/235910@main