Make sure the Frame always gets destroyed on the main thread https://bugs.webkit.org/show_bug.cgi?id=224213 <rdar://75131946>
Reviewed by Alex Christensen.

Make sure the Frame always gets destroyed on the main thread.

Since http://trac.webkit.org/changeset/181791, Frame is ThreadSafeRefCounted and captured
on other threads. We should make sure the Frame is always destroyed on the main thread
though, since it is a main thread object and we hold WeakPtrs to it in many places.

* page/AbstractFrame.h:


Canonical link: https://commits.webkit.org/236133@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275475 268f45cc-cd09-0410-ab3c-d52691b4dbfc