CloneDeserializer should use ArrayBuffer::tryCreate https://bugs.webkit.org/show_bug.cgi?id=224218
Reviewed by Antti Koivisto.

Source/WebCore:

Right now CloneDeserializer assumes that every ArrayBuffer allocation during
deserialization will succeed. This is silly since it's an array-like object.
It should call tryCreate and fail the deserialization instead.

Test: fast/dom/Window/post-message-large-array-buffer-should-not-crash.html

* bindings/js/SerializedScriptValue.cpp:
(WebCore::CloneDeserializer::readArrayBuffer):

LayoutTests:

This test was generated by a fuzzer so it allocates a large Array backing store
by doing Object.defineProperty on a large offset. That said, I chose to leave it
because it's sometimes useful to do things in different ways for testing.

Also, skip the test on windows because we seem to throw a stack overflow error.
Not sure why this happens but it's not super important that this particular
test runs on all ports as we're mostly trying to just unblock the fuzzer.

* fast/dom/Window/post-message-large-array-buffer-should-not-crash-expected.txt: Added.
* fast/dom/Window/post-message-large-array-buffer-should-not-crash.html: Added.
* platform/win/TestExpectations:


Canonical link: https://commits.webkit.org/236165@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275508 268f45cc-cd09-0410-ab3c-d52691b4dbfc