Web Inspector: Grid overlay areas appear offset from their actual areas when the grid container has a border set https://bugs.webkit.org/show_bug.cgi?id=224201
Reviewed by BJ Burg.

As of r274096, it is no longer correct to subtract the first track's position for either axis from the edge
lines of areas. This change resolves that, bringing this math in line with how other grid lines are laid out.

* inspector/InspectorOverlay.cpp:
(WebCore::InspectorOverlay::buildGridOverlay):


Canonical link: https://commits.webkit.org/236185@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275529 268f45cc-cd09-0410-ab3c-d52691b4dbfc