Unreviewed, fix the internal build after r275640 https://bugs.webkit.org/show_bug.cgi?id=224239
* WebKit.xcodeproj/project.pbxproj:
This source file was added to the project as a compilation target, but it also appears in unified sources; this
causes the `encode` and `decode` methods to be defined twice, leading to a linker error. Fix this by only
including it as a unified source.


Canonical link: https://commits.webkit.org/236299@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275663 268f45cc-cd09-0410-ab3c-d52691b4dbfc