[css-grid] Fix min/max widths of grid affected by ancestor https://bugs.webkit.org/show_bug.cgi?id=222100
Patch by Ziran Sun <zsun@igalia.com> on 2021-04-09
LayoutTests/imported/w3c:

Reviewed by Javier Fernandez.

* web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-002-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-lr-002-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-rl-002-expected.txt:

Source/WebCore:

Reviewed by Reviewed by Javier Fernandez.

It's a reland of r273435, which got reverted because it broke
imported/w3c/web-platform-tests/css/css-flexbox/percentage-max-height-003.html. This change
narrows down the fix specificallly to the case when logical-width is recomputed.
We need to recalculate min/max widths of child that depend on the ancestor.
Before update logical-width, for element that needs preferredWidth recalcution,
it is necessary to make sure that min/max widths are set dirty.

This change is an import of chromium CL at
https://chromium-review.googlesource.com/c/chromium/src/+/527640/
Only the parts that apply to this issue are imported.

Tests were already imported in WPT.

* rendering/RenderBlock.cpp:
(WebCore::shouldRecalculateMinMaxWidthsAffectedByAncestor):
(WebCore::RenderBlock::recomputeLogicalWidth):

Canonical link: https://commits.webkit.org/236332@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275754 268f45cc-cd09-0410-ab3c-d52691b4dbfc