border-image-width computed values should be a calc() value if it contains a percentage https://bugs.webkit.org/show_bug.cgi?id=224420
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

We now pass the final test that was a FAIL result for the border-image-width computed style test.
We also pass a fair few border-image-width interpolation tests since we used to fail them as a
result of the expected value being expected to be a resolved value rather than a calc() value
due the computed style we returned for such values. There are a few PASS results than became
FAIL results because we now correctly expect a calc() value, but we return the keyframe value
as-is for 0 and 1 progress, a bug we will fix shortly.

* web-platform-tests/css/css-backgrounds/animations/border-image-width-interpolation-expected.txt:
* web-platform-tests/css/css-backgrounds/parsing/border-image-width-computed-expected.txt:

Source/WebCore:

When handling calc() values containing a percentage, we must retain the calc() form when creating
the computed style for border-image-width.

* css/CSSToStyleMap.cpp:
(WebCore::CSSToStyleMap::mapNinePieceImageQuad):


Canonical link: https://commits.webkit.org/236403@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275834 268f45cc-cd09-0410-ab3c-d52691b4dbfc