[GTK4] WTR crash in PlatformWebView::dismissAllPopupMenus() https://bugs.webkit.org/show_bug.cgi?id=224487
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2021-04-13
Reviewed by Philippe Normand.

This is because we are using a for loop to iterate children that are destroyed. We should not be destroying the
children, though, we should just hide them, since they are owned by the web view.

* WebKitTestRunner/gtk/PlatformWebViewGtk.cpp:
(WTR::PlatformWebView::dismissAllPopupMenus):

Canonical link: https://commits.webkit.org/236450@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275885 268f45cc-cd09-0410-ab3c-d52691b4dbfc