Use a JSString for the TerminationException value instead of a Symbol. https://bugs.webkit.org/show_bug.cgi?id=224490
Reviewed by Yusuke Suzuki.

This makes it convertible to a String for clients that wish to report the exception.
Clients, in this case, does not apply to JS code, only C++ code that catches the
exception at the outermost point to handle the termination.  The TerminationException
value is not visible to JS code because the TerminationException cannot be caught.
So, this change is transparent to JS code.

* runtime/VM.cpp:
(JSC::VM::ensureTerminationException):


Canonical link: https://commits.webkit.org/236467@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@275902 268f45cc-cd09-0410-ab3c-d52691b4dbfc